+++ src/widgets/ColorButton.cpp
@@ -19,9 +19,12 @@
 
 #include "ColorButton.h"
 
-ColorButton::ColorButton(QWidget* parent) : QPushButton(parent)
+ColorButton::ColorButton(QWidget* parent) :
+	QPushButton(parent),
+	mShowAlphaChannel(false)
 {
     connect(this, &ColorButton::clicked, this, &ColorButton::openDialog);
+	setIconSize(QSize(48, iconSize().height()));
 }
 
 QColor ColorButton::color() const
@@ -34,20 +37,46 @@ void ColorButton::setColor(const QColor& color)
     mColor = color;
     auto pixmapIcon = createPixmapFromColor(mColor);
     setIcon(pixmapIcon);
+    setToolTip(mShowAlphaChannel ? mColor.name(QColor::HexArgb) : mColor.name());
+}
+
+void ColorButton::setShowAlphaChannel(bool enabled)
+{
+	mShowAlphaChannel = enabled;
 }
 
 QPixmap ColorButton::createPixmapFromColor(const QColor& color)
 {
-    QPixmap pixmap(32, 32);
-    pixmap.fill(color);
-    return pixmap;
+	QImage tiledBackground(QSize(10, 10), QImage::Format_ARGB32_Premultiplied);
+	tiledBackground.fill(Qt::white);
+	QPainter tiledBackgroundPainter(&tiledBackground);
+	tiledBackgroundPainter.setPen(Qt::NoPen);
+	tiledBackgroundPainter.setBrush(Qt::gray);
+	tiledBackgroundPainter.drawRect(0, 0, 5, 5);
+	tiledBackgroundPainter.drawRect(5, 5, 10, 10);
+
+	QPixmap pixmap(iconSize());
+	QPainter pixmapPainter(&pixmap);
+	pixmapPainter.setPen(Qt::gray);
+	pixmapPainter.setBrush(tiledBackground);
+	pixmapPainter.drawRect(0,0, iconSize().width() - 1, iconSize().height() - 1);
+	pixmapPainter.setBrush(color);
+	pixmapPainter.drawRect(0,0, iconSize().width() - 1, iconSize().height() - 1);
+	return pixmap;
 }
 
 void ColorButton::openDialog()
 {
-    auto color = QColorDialog::getColor(mColor, parentWidget());
-    if (!color.isValid()) {
-        return;
-    }
-    setColor(color);
+	QFlags<QColorDialog::ColorDialogOption> options;
+
+	if(mShowAlphaChannel) {
+		options |= QColorDialog::ShowAlphaChannel;
+	}
+
+	auto color = QColorDialog::getColor(mColor, parentWidget(), QString(), options);
+	
+	if (mShowAlphaChannel) 
+	if (color.isValid() && color != mColor) {
+		setColor(color);
+	}
 }